Fixed Bugs in Multiplication and Numerical Double Check
Fixed bugs in QuadarticFieldTower::Product so that precision
multiplication is now working properly. Also, fixed bugs involving
conversion of roots and coordinates to std::complex<double>. Also,
changed from use of std::complex<float> to std::complex<double>.

+++ QuadraticField.h
@@ -24,7 +24,7 @@ public:
 	int GetSqrtP(){return sqrtp;}
 	int GetSqrtQ(){return sqrtq;}
 	Rational GetSqrt(){return Rational(sqrtp, sqrtq);}
-	float ToFloat();
+	double ToFloat();
 
 private:
 
@@ -70,12 +70,12 @@ class QuadraticFieldTower {
 		std::string Print();
 		std::string PrintRootList();
 		std::string PrintCoords(std::vector<Rational> & coords);
-		std::complex<float> CoordsToComplex(std::vector<Rational> & coords);
+		std::complex<double> CoordsToComplex(std::vector<Rational> & coords);
 
 	private:
 		int degree, numsquares;
 		std::vector< std::vector<Rational> > squares;
-		std::vector< std::complex<float> > complexroots;
+		std::vector< std::complex<double> > complexroots;
 };
 //////////////////////////////////
 class QuadraticField {
